Updating 404 Errors on logout and login
