+++ logout.php
@@ -4,7 +4,7 @@ session_start();
 session_destroy();
 
 // Redirect to login page
-header('Location: /Login/User/index.php');
+header('Location: ./Login/User/index.php');
 exit(0);
 
 ?>
